+++ src/gei/Commands/MigrateRepoCommand.cs
@@ -108,7 +108,9 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
         };
         public Option<string> AwsRegion { get; } = new("--aws-region")
         {
-            Description = "If using AWS, the AWS region. If not provided, it will be read from AWS_REGION environment variable. Defaults to us-east-1."
+            Description = "If using AWS, the AWS region. If not provided, it will be read from AWS_REGION environment variable. " +
+                          "Defaults to us-east-1 if neither the argument nor the environment variable is set. " +
+                          "In a future release, you will be required to set an AWS region if using AWS S3 as your blob storage provider."
         };
         public Option<bool> NoSslVerify { get; } = new("--no-ssl-verify")
         {